+++ djangae/tests/test_db_fields.py
@@ -65,6 +65,24 @@ class ModelWithManyCounters(models.Model):
         app_label = "djangae"
 
 
+def update_denormalized_counter(instance, step, is_reset=False):
+    instance.denormalized_counter += step
+
+    if is_reset:
+        instance.reset_count += 1
+
+    instance.save()
+
+
+class ModelWithCountersWithCallback(models.Model):
+    counter = ShardedCounterField(on_change=update_denormalized_counter)
+    denormalized_counter = models.IntegerField(default=0)
+    reset_count = models.IntegerField(default=0)
+
+    class Meta:
+        app_label = "djangae"
+
+
 class ModelWithCounterWithManyShards(models.Model):
     # The DEFAULT_SHARD_COUNT is based on the max allowed in a Datastore transaction
     counter = ShardedCounterField(shard_count=DEFAULT_SHARD_COUNT+5)
@@ -280,6 +298,30 @@ class ShardedCounterTest(TestCase):
         self.assertEqual(instance.counter.shard_count(), 0)
         instance.counter.populate()
         self.assertEqual(instance.counter.shard_count(), 24)  # The default shard count is 24
+
+    def test_on_change_callback_called(self):
+        """ Test that if ShardedCounter has on_change callback specified, it
+            will be called after any change to the counter.
+        """
+        instance = ModelWithCountersWithCallback.objects.create()
+        self.assertEquals(instance.denormalized_counter, 0)
+
+        # callback called after increment
+        instance.counter.increment(5)
+        instance.refresh_from_db()
+        self.assertEquals(instance.denormalized_counter, 5)
+
+        # after decrement
+        instance.counter.decrement(1)
+        instance.refresh_from_db()
+        self.assertEquals(instance.denormalized_counter, 4)
+
+        # and reset
+        self.assertEquals(instance.reset_count, 0)
+        instance.counter.reset()
+        instance.refresh_from_db()
+        self.assertEquals(instance.denormalized_counter, 0)
+        self.assertEquals(instance.reset_count, 1)
 
 
 class IterableFieldTests(TestCase):